refactor motor interface

+++ node2/motor.h
@@ -4,18 +4,9 @@
 
 void motor_init();
 
-void motor_command_speed(int16_t speed);
+void motor_command_speed(uint8_t speed);
 
-/**
- * @brief Command motor position.
- *
- * Range: [-100, 100], where -100 is all the way to
- * the left; 100 is all the way to the right, and 0
- * is the neutral position.
- *
- * @param position Commanded position.
- */
-void motor_command_position(int16_t position);
+void motor_command_position(uint8_t position);
 
 uint16_t motor_encoder_read();
